Refactor pool creation and input handling
- Update denomination type from Float to Double for increased precision
- Implement debug logging for PSBT and transaction data
- Enhance input validation for pool creation and joining
- Refine error handling in HTTP client
- Adjust UI elements for better visibility
- Optimize BigDecimal calculations in RegisterInputViewModel

+++ composeApp/src/commonMain/kotlin/invincible/privacy/joinstr/model/Credentials.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 data class Credentials(
     val id: String,
     val publicKey: String,
-    val denomination: Float,
+    val denomination: Double,
     val peers: Int,
     val timeout: Long,
     val relay: String,
